Refactored code to update lastLogMark only when replaying journal and addressed nits
